backend/requirements.txt
# FastAPI & Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Supabase
supabase==2.3.4
postgrest==0.15.0

# Celery & Redis
celery==5.3.6
redis==5.0.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0

# HTTP Clients
httpx<0.26,>=0.24
aiohttp==3.9.1

# Document Processing
python-docx==1.1.0
python-magic==0.4.27

# Audio Processing
pydub==0.25.1

# AI/ML
openai==1.10.0
tiktoken==0.5.2

# Google APIs
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Linear
gql[all]==3.5.0

# File Storage
boto3==1.34.34  # For S3-compatible storage

# Utilities
python-dateutil==2.8.2
pytz==2024.1
tenacity==8.2.3

# CLI & Progress
rich==13.7.0
click==8.1.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.6.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Logging & Monitoring
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0
